# PERF: 10x speedup in Series/DataFrame construction for lists of ints