@jreback 2019-01-05
actually there are a few more cases where the same fix could be applied