@qwhelan 2019-01-05
@jreback Just to clarify - the speedup comes from preventing an object-to-object comparison for `val > oUINT_MAX` and the like (which is avoided in `saw_int(int(val))` by casting the input). I think this is the only change needed for that precise case, but happy to implement any that I'm missing here.